Reorganized folder structures. Still not sure about it.
